# refactor: 优化交叉表渲染性能