--- a/packages/s2-core/src/common/constant/basic.ts
+++ b/packages/s2-core/src/common/constant/basic.ts
@@ -77,3 +77,6 @@ export const IMAGE = 'image';
 
 // 角头最大占整个容器的比例 (0-1)
 export const CORNER_MAX_WIDTH_RATIO = 0.5;
+
+/** 布局采样数 */
+export const LAYOUT_SAMPLE_COUNT = 50;