### 👀 PR includes

<!-- Add completed items in this PR, and change [ ] to [x]. -->

🎨 Enhance

- [x] Improve the performance

### 📝 Description

#### 1. 树状结构行头布局排序效率
修改：buildRowTreeHierarchy、getListBySorted
影响：tree 的行头布局

##### 之前
同一层级节点 layout 时，都会全量处理 sortedDimensionValues 的值。
先做筛选、再去掉 `前一层[&]` 的前缀，如有 ['四川[&]成都', '浙江[&]杭州'] -> ['四川[&]成都'] -> ['成都']

##### 现在
无需要做过滤和去前缀，转而在 getListBySorted 做比较时，把维度值 dimValues 补齐，如 `成都` -> `四川[&]成都`。
因为 sortedDimensionValues 数组比较大，包含了同一层级的所有维度值。而 dimValues 只是同级的小分支，数据个数比较少。
且 dimValues 只有 1 个时（单节点），不用调用 sortFunc，进一步节约执行时间。

#### 2. compact 模式列宽采样计算
修改：grid、tree 的 compact 模式布局

##### 之前
使用 getMultiData 取了当前列的所有数据，再 slice 50 个。

##### 现在
直接循环 50 次，按行头叶子节点与当前列的交叉，执行获取50次数据进行采样。



#### 3. sortedDimensionValues 的初始化效率
修改：transformIndexesData、transformIndexesData

##### 之前
在 transformDimensionsValues 遍历所有数据时，反复检查当前维度值是否插入了 sortedDimensionValues，没有就插入。
其中检查函数 'includes' 执行很多次，比较耗时。

##### 现在
transformDimensionsValues 不再包含副作用，仅按配置的维度，返回维度数据。
因为 getDataPath 会遍历生成 pivotMap，所以在遇到创建新节点时构造 sortedDimensionValues 即可，不用检查重复。


### 🖼️ Screenshot

| Before | After |
| ------ | ----- |
| 约 20s | 约 360ms |
|  ![image](https://user-images.githubusercontent.com/6716092/173274383-72f43a55-f701-45f7-9cb6-662dc77f610e.png)     |  ![image](https://user-images.githubusercontent.com/6716092/173274406-c3554c03-9ced-4650-a0e4-f0c6310b8447.png)   |

### 🔗 Related issue link

<!-- close #0 -->

### 🔍 Self Check before Merge

- [ ] Add or update relevant Docs.
- [ ] Add or update relevant Demos.
- [ ] Add or update relevant TypeScript definitions.