@lijinke666 2022-06-17
:tada: This PR is included in version @antv/s2-v1.20.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/@antv/s2/v/1.20.0)
- [GitHub release](https://github.com/antvis/S2/releases/tag/%40antv/s2-v1.20.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: