@lijinke666 2022-11-25
:tada: This PR is included in version 1.0.0 :tada:

The release is available on [GitHub release](https://github.com/antvis/S2/releases/tag/v1.0.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: